Updated data type for itemId and refId
* Updated refId data type from int to long
* Updated itemId data type from int to long
* Updated target framework from Net46 to Net48
* Updated dependencies to Net48

+++ Source/Podio .NET/Models/Grant.cs
@@ -12,7 +12,7 @@ namespace PodioAPI.Models
         public string refType { get; set; }
 
         [JsonProperty("ref_id")]
-        public int refId { get; set; }
+        public long refId { get; set; }
 
         [JsonProperty("action")]
         public string Action { get; set; }
