+++ Source/Podio .NET/Services/BatchService.cs
@@ -44,7 +44,7 @@ namespace PodioAPI.Services
         /// <param name="refId"></param>
         /// <param name="plugin">The plugin can either be "app_import", "app_export", "space_contact_import" or "app_content".</param>
         /// <returns></returns>
-        public async Task<List<Batch>> GetRunningBatches(string refType, int refId, string plugin)
+        public async Task<List<Batch>> GetRunningBatches(string refType, long refId, string plugin)
         {
             string url = string.Format("/batch/{0}/{1}/{2}/running/", refType, refId, plugin);
             return await _podio.Get<List<Batch>>(url);
